auth-service/requirements.txt
# Core
fastapi>=0.110,<1
uvicorn[standard]>=0.30,<1

# Database (async)
SQLAlchemy[asyncio]>=2.0,<3
asyncpg>=0.29,<1
psycopg>=3.1,<4

# Validation
pydantic>=2.7,<3
pydantic-settings>=2.2,<3
email-validator>=2.1,<3

# Auth & Security (versiones específicas compatibles)
python-jose[cryptography]>=3.3,<4
passlib>=1.7,<2
bcrypt==4.0.1

# Utils
python-dotenv>=1.0,<2
python-multipart>=0.0.6,<1